--- demos/training/application/Dockerfile.inference
@@ -1,11 +1,0 @@
-# The cbl-mariner image is stuck at python 3.9. Pulling the image from the  docker mirror till
-# mariner bumps up the python version.
-# FROM mcr.microsoft.com/cbl-mariner/base/python:3
-FROM mcr.microsoft.com/mirror/docker/library/python:3.11
-WORKDIR /code
-COPY ./requirements.txt /code/requirements.txt
-RUN pip install --no-cache-dir --upgrade -r /code/requirements.txt
-
-COPY ./host_model.py app/
-ENV APPLICATION_PORT=80
-CMD ["python3", "./app/host_model.py"]
