--- demos/training/application/PyTorchTraining.py
@@ -1,222 +1,0 @@
-# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
-
-from torchvision.datasets import CIFAR10
-from torchvision.transforms import transforms
-from torch.utils.data import DataLoader
-
-import torch
-import torch.nn as nn
-import torchvision
-import torch.nn.functional as F
-
-from torch.optim import Adam
-
-# Loading and normalizing the data.
-# Define transformations for the training and test sets
-transformations = transforms.Compose([
-    transforms.ToTensor(),
-    transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
-])
-
-# CIFAR10 dataset consists of 50K training images. We define the batch size of 10 to load 5,000 batches of images.
-batch_size = 10
-number_of_labels = 10 
-
-# Create an instance for training. 
-# When we run this code for the first time, the CIFAR10 train dataset will be downloaded locally. 
-train_set =CIFAR10(root="./data",train=True,transform=transformations,download=True)
-
-# Create a loader for the training set which will read the data within batch size and put into memory.
-train_loader = DataLoader(train_set, batch_size=batch_size, shuffle=True, num_workers=0)
-print("The number of images in a training set is: ", len(train_loader)*batch_size)
-
-# Create an instance for testing, note that train is set to False.
-# When we run this code for the first time, the CIFAR10 test dataset will be downloaded locally. 
-test_set = CIFAR10(root="./data", train=False, transform=transformations, download=True)
-
-# Create a loader for the test set which will read the data within batch size and put into memory. 
-# Note that each shuffle is set to false for the test loader.
-test_loader = DataLoader(test_set, batch_size=batch_size, shuffle=False, num_workers=0)
-print("The number of images in a test set is: ", len(test_loader)*batch_size)
-
-print("The number of batches per epoch is: ", len(train_loader))
-classes = ('plane', 'car', 'bird', 'cat', 'deer', 'dog', 'frog', 'horse', 'ship', 'truck')
-
-# Define a convolution neural network
-class Network(nn.Module):
-    def __init__(self):
-        super(Network, self).__init__()
-        
-        self.conv1 = nn.Conv2d(in_channels=3, out_channels=12, kernel_size=5, stride=1, padding=1)
-        self.bn1 = nn.BatchNorm2d(12)
-        self.conv2 = nn.Conv2d(in_channels=12, out_channels=12, kernel_size=5, stride=1, padding=1)
-        self.bn2 = nn.BatchNorm2d(12)
-        self.pool = nn.MaxPool2d(2,2)
-        self.conv4 = nn.Conv2d(in_channels=12, out_channels=24, kernel_size=5, stride=1, padding=1)
-        self.bn4 = nn.BatchNorm2d(24)
-        self.conv5 = nn.Conv2d(in_channels=24, out_channels=24, kernel_size=5, stride=1, padding=1)
-        self.bn5 = nn.BatchNorm2d(24)
-        self.fc1 = nn.Linear(24*10*10, 10)
-
-    def forward(self, input):
-        output = F.relu(self.bn1(self.conv1(input)))      
-        output = F.relu(self.bn2(self.conv2(output)))     
-        output = self.pool(output)                        
-        output = F.relu(self.bn4(self.conv4(output)))     
-        output = F.relu(self.bn5(self.conv5(output)))     
-        output = output.view(-1, 24*10*10)
-        output = self.fc1(output)
-
-        return output
-
-from torch.autograd import Variable
-
-# Function to test the model with the test dataset and print the accuracy for the test images
-def testAccuracy(model):
-    
-    model.eval()
-    accuracy = 0.0
-    total = 0.0
-    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-    
-    with torch.no_grad():
-        for data in test_loader:
-            images, labels = data
-            # run the model on the test set to predict labels
-            outputs = model(images.to(device))
-            # the label with the highest energy will be our prediction
-            _, predicted = torch.max(outputs.data, 1)
-            total += labels.size(0)
-            accuracy += (predicted == labels.to(device)).sum().item()
-    
-    # compute the accuracy over all test images
-    accuracy = (100 * accuracy / total)
-    return(accuracy)
-
-
-# Training function. We simply have to loop over our data iterator and feed the inputs to the network and optimize.
-def train(model, device, epoch):
-    # Define the loss function with Classification Cross-Entropy loss and an optimizer with Adam optimizer
-    loss_fn = nn.CrossEntropyLoss()
-    optimizer = Adam(model.parameters(), lr=0.001, weight_decay=0.0001)
-    running_loss = 0.0
-
-    for i, (images, labels) in enumerate(train_loader, 0):
-        
-        # get the inputs
-        images = Variable(images.to(device))
-        labels = Variable(labels.to(device))
-
-        # zero the parameter gradients
-        optimizer.zero_grad()
-        # predict classes using images from the training set
-        outputs = model(images)
-        # compute the loss based on model output and real labels
-        loss = loss_fn(outputs, labels)
-        # backpropagate the loss
-        loss.backward()
-        # adjust parameters based on the calculated gradients
-        optimizer.step()
-
-        # Let's print statistics for every 1,000 images
-        running_loss += loss.item()     # extract the loss value
-        if i % 1000 == 999:
-            # print every 1000 (twice per epoch)
-            print('[%d, %5d] loss: %.3f' %
-                    (epoch + 1, i + 1, running_loss / 1000))
-            # zero the loss
-            running_loss = 0.0
-
-    # Compute and print the average accuracy for this epoch when tested over all 10000 test images
-    accuracy = testAccuracy(model)
-    print('For epoch', epoch+1,'the test accuracy over the whole test set is %d %%' % (accuracy))
-
-
-import matplotlib.pyplot as plt
-import numpy as np
-import torch.onnx 
-
-#Function to Convert to ONNX 
-def Convert_ONNX(): 
-
-    # set the model to inference mode 
-    model.eval() 
-
-    # Let's create a dummy input tensor  
-    dummy_input = torch.randn(1, 3, 32, 32, requires_grad=True)  
-
-    # Export the model   
-    torch.onnx.export(model,         # model being run 
-         dummy_input,       # model input (or a tuple for multiple inputs) 
-         "ImageClassifier.onnx",       # where to save the model  
-         export_params=True,  # store the trained parameter weights inside the model file 
-         opset_version=10,    # the ONNX version to export the model to 
-         do_constant_folding=True,  # whether to execute constant folding for optimization 
-         input_names = ['modelInput'],   # the model's input names 
-         output_names = ['modelOutput'], # the model's output names 
-         dynamic_axes={'modelInput' : {0 : 'batch_size'},    # variable length axes 
-                                'modelOutput' : {0 : 'batch_size'}}) 
-
-    print(" ") 
-    print('Model has been converted to ONNX') 
-
-import time
-import os
-
-def buildAndTrain(path):
-
-    # Define your execution device
-    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-    print("The model will be running on", device, "device")
-
-    # Instantiate a neural network model
-    model = Network()
-
-    if (os.path.exists(path)):
-        print("Model exists")
-
-        # Load the state dict
-        model.load_state_dict(torch.load(path))
-
-    else:
-        print("Initializing the model")
-
-    # Convert model parameters and buffers to CPU or Cuda
-    model.to(device)
-
-    # Train
-    for epoch in range(5):  # loop over the dataset multiple times
-        train(model, device, epoch)
-
-    # Save the model
-    torch.save(model.state_dict(), path)
-
-
-if __name__ == "__main__":
-
-    start = time.time()
-    path = "myFirstModel.pth"
-
-    # Let's build & save our model
-    #train(5)
-    buildAndTrain(path)
-
-    print('Finished Training')
-
-    # <code to time>
-    end = time.time()
-    print(f"Time taken to run was {end-start} seconds")
-
-    # Test which classes performed well
-    #testAccuracy()
-    
-    # Let's load the model we just created and test the accuracy per label
-    model = Network()
-
-    model.load_state_dict(torch.load(path))
-
-    # Test with batch of images
-    #testBatch()
-
-    # Conversion to ONNX 
-    Convert_ONNX()
