+++ demos/training/datasink/fabrikam/output/.gitignore
@@ -1,0 +1,1 @@
+*/*
