+++ demos/training/datasource/contosso/data/.gitignore
@@ -1,1 +1,1 @@
-*/*
+*
