+++ demos/training/datasource/fabrikam/model/.gitignore
@@ -1,1 +1,1 @@
-*/*
+*.pth
