+++ demos/training/model/get_model.py
@@ -1,32 +1,55 @@
 import logging
-from optimum.onnxruntime import ORTModelForSequenceClassification
-from transformers import DistilBertForSequenceClassification, DistilBertTokenizerFast
+import torch
 
 from pydantic_settings import BaseSettings
 from pydantic import Field
 
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger("get_model")
+import torch.nn as nn
+import torch.nn.functional as F
 
+# Define a convolution neural network
+class Network(nn.Module):
+    def __init__(self):
+        super(Network, self).__init__()
+
+        self.conv1 = nn.Conv2d(in_channels=3, out_channels=12, kernel_size=5, stride=1, padding=1)
+        self.bn1 = nn.BatchNorm2d(12)
+        self.conv2 = nn.Conv2d(in_channels=12, out_channels=12, kernel_size=5, stride=1, padding=1)
+        self.bn2 = nn.BatchNorm2d(12)
+        self.pool = nn.MaxPool2d(2,2)
+        self.conv4 = nn.Conv2d(in_channels=12, out_channels=24, kernel_size=5, stride=1, padding=1)
+        self.bn4 = nn.BatchNorm2d(24)
+        self.conv5 = nn.Conv2d(in_channels=24, out_channels=24, kernel_size=5, stride=1, padding=1)
+        self.bn5 = nn.BatchNorm2d(24)
+        self.fc1 = nn.Linear(24*10*10, 10)
+
+    def forward(self, input):
+        output = F.relu(self.bn1(self.conv1(input)))
+        output = F.relu(self.bn2(self.conv2(output)))
+        output = self.pool(output)
+        output = F.relu(self.bn4(self.conv4(output)))
+        output = F.relu(self.bn5(self.conv5(output)))
+        output = output.view(-1, 24*10*10)
+        output = self.fc1(output)
+
+        return output
 
 class AppSettings(BaseSettings, cli_parse_args=True):
     output_path: str = Field(alias="output-path")
 
-
 settings = AppSettings()
-model_id = "distilbert/distilbert-base-uncased"
 
-logging.info(f"Fetching model {model_id} from huggingface Hub.")
-model = DistilBertForSequenceClassification.from_pretrained(model_id)
-logging.info(f"Saving model to path {settings.output_path}")
-model.save_pretrained(f"{settings.output_path}/pytorch")
+# Generating an empty model to be trained.
 
-tokenizer = DistilBertTokenizerFast.from_pretrained(model_id)
-logging.info(f"Saving tokenizer to path {settings.output_path}")
-tokenizer.save_pretrained(f"{settings.output_path}/pytorch")
+#device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+#logging.info(f"The model will be running on", device, "device")
 
-model = ORTModelForSequenceClassification.from_pretrained(model_id, export=True)
-logging.info(f"Saving ONNX model to path {settings.output_path}")
-model.save_pretrained(f"{settings.output_path}/onnx")
+# Instantiate a neural network model
+model = Network()
 
-# TODO: Should we tune the model to give better inference results?
+# Convert model parameters and buffers to CPU or Cuda
+#model.to(device)
+
+# Save the model
+torch.save(model.state_dict(), f"{settings.output_path}/model.pth")
+logging.info(f"Model saved to path {settings.output_path}/model.pth")
